+++ examples/02-spritesheets/assets.lua
@@ -1,0 +1,2 @@
+rbxassetid://4643741386
+rbxassetid://4643741424