+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 	* This target always fails to upload, and is useful to verify that all assets have been uploaded as part of a CI job.
 * Added `asset-cache-path` config option. If specified, Tarmac will download managed assets from Roblox.com to populate the given directory with.
 * Added `upload-to-group-id` config option to require that all uploaded assets are uploaded to the given group.
+* Added `asset-list-path` for generating a list of all asset URLs referred to by the Tarmac project.
+	* This output format is intended for consumption by other tools.
 
 ## 0.4.0 (2020-03-04)
 * Tarmac now "alpha-bleeds" packed image spritesheets to prevent artifacts from appearing when resized in Roblox.